requirements.txt
python-telegram-bot==20.5
telethon==1.31.1
python-dotenv==1.0.0
pymongo==4.5.0
fuzzywuzzy==0.18.0
python-Levenshtein==0.22.0 